requirements/dev.txt
Django==2.2.6
djangorestframework==3.10.3
django_configurations==2.1
django_dotenv==1.4.2
